Add a WAITING state to the Source FSM.
This is required to correctly resolve sources' initialize promises. If a source was initialized
but did not exist, it never wopuld have resolved its promise. Instead, on the first update,
the state transitions away from INITIALIZING even if the data is NO_EXIST, ensuring that
the promise always resolves for a healthy source.